aider: Added a Rollup configuration to build the `index.js` file.
# Aider chat conversation:

USER: add a rollup config to build the index.js

ASSISTANT: To add a Rollup configuration to build the `index.js` file, we need to create a new file named `rollup.config.js` in the root directory. This file will contain the configuration for Rollup to bundle your `index.js` file.

Here is the edit block for the new `rollup.config.js` file:

```javascript
rollup.config.js
<<<<<<< ORIGINAL

=======
const path = require('path');

module.exports = {
  input: path.resolve(__dirname, 'index.js'),
  output: {
    file: path.resolve(__dirname, 'dist/bundle.js'),
    format: 'cjs',
  },
};
>>>>>>> UPDATED
```

This configuration tells Rollup to take `index.js` as the input file and output the bundled code to `dist/bundle.js` in CommonJS format. The `path.resolve` function is used to ensure that the paths are correct regardless of the operating system.

Please note that you will need to install Rollup and add it to your `package.json` file. If you need help with this, please let me know.