aider: Added `build` script and `rollup` as a devDependency in package.json.
# Aider chat conversation:

USER: I added these *read-write* files: package.json

ASSISTANT: Great! Now that we have the `package.json` file in read-write mode, we can add the necessary scripts and dependencies for Rollup.

We need to add a `build` script that will use Rollup to bundle our `index.js` file. We also need to add `rollup` as a devDependency.

Here is the edit block for the `package.json` file:

```json
package.json
<<<<<<< ORIGINAL
{
  "name": "@ossph/opensource",
  "version": "1.0.1",
  "description": "Generate Open Source Project",
  "main": "index.js",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1",
    "publish": "npm publish --access public"
  },
  "repository": {
    "type": "git",
    "url": "git+ssh://git@github.com/OSSPhilippines/opensource.git"
  },
  "keywords": [
    "opensource",
    "oss",
    "generator"
  ],
  "author": "Joff Tiquez",
  "license": "MIT",
  "bugs": {
    "url": "https://github.com/OSSPhilippines/opensource/issues"
  },
  "homepage": "https://github.com/OSSPhilippines/opensource#readme"
}
=======
{
  "name": "@ossph/opensource",
  "version": "1.0.1",
  "description": "Generate Open Source Project",
  "main": "index.js",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1",
    "publish": "npm publish --access public",
    "build": "rollup -c"
  },
  "repository": {
    "type": "git",
    "url": "git+ssh://git@github.com/OSSPhilippines/opensource.git"
  },
  "keywords": [
    "opensource",
    "oss",
    "generator"
  ],
  "author": "Joff Tiquez",
  "license": "MIT",
  "bugs": {
    "url": "https://github.com/OSSPhilippines/opensource/issues"
  },
  "homepage": "https://github.com/OSSPhilippines/opensource#readme",
  "devDependencies": {
    "rollup": "^2.56.3"
  }
}
>>>>>>> UPDATED
```

This configuration adds a `build` script that runs `rollup -c`, which will use the configuration in `rollup.config.js` to bundle your code. It also adds `rollup` as a devDependency with a version of `^2.56.3`. You can adjust the version as necessary.

Please note that you will need to run `npm install` after making these changes to install Rollup.