aider: Added `node_modules/` to the `.gitignore` file to ignore the `node_modules` directory.
# Aider chat conversation:

USER: ignore node_modules

ASSISTANT: To ignore the `node_modules` directory, we need to add it to the `.gitignore` file. However, the `.gitignore` file is currently read-only. Could you please add it to the chat in read-write mode?

USER: I added these *read-write* files: .gitignore

ASSISTANT: Great! Now that we have the `.gitignore` file in read-write mode, we can add `node_modules/` to it. This will tell Git to ignore the `node_modules` directory.

Here is the edit block for the `.gitignore` file:

```gitignore
.gitignore
<<<<<<< ORIGINAL
.aider*
=======
.aider*
node_modules/
>>>>>>> UPDATED
```

This configuration adds `node_modules/` to the `.gitignore` file, which will prevent Git from tracking the `node_modules` directory.