aider: Added Jest test for `index.js` that mocks the `fs` module and asserts that `writeFileSync` is called with the correct arguments.
# Aider chat conversation:

USER: write a jest test for `index.js`

ASSISTANT: To write a Jest test for `index.js`, we need to create a new file named `index.test.js` in the same directory. This file will contain our test cases.

Here's an example of how we can structure our test:

```javascript
index.test.js
<<<<<<< ORIGINAL

=======
const fs = require('fs');
const path = require('path');
jest.mock('fs');

describe('index.js tests', () => {
  afterEach(() => {
    jest.clearAllMocks();
  });

  test('should write README.md with correct content', () => {
    const packageJson = {
      name: 'package name',
      description: 'package description',
      author: 'package author',
      license: 'package license',
    };
    fs.readFileSync.mockReturnValue(JSON.stringify(packageJson));
    require('./index.js');
    expect(fs.writeFileSync).toHaveBeenCalledWith('README.md', `# ${packageJson.name}\n\n${packageJson.description}\n\n## Author\n\n${packageJson.author}\n\n## License\n\n${packageJson.license}`);
  });
});
>>>>>>> UPDATED
```

This test case mocks the `fs` module and its `readFileSync` and `writeFileSync` methods. It then requires the `index.js` file, which triggers the execution of the script. Finally, it asserts that `writeFileSync` was called with the correct arguments, which are the filename and the content of the README.md file.

Please add this test file to your project and run the tests with Jest.