aider: Generated a LICENSE file for ISC in src/templates/licenses.
# Aider chat conversation:

USER: in src/templates/licences generate a LICENSE file for ISC

ASSISTANT: Sure, I will create a new file `src/templates/licenses/ISC_LICENSE.md` with the standard ISC license text. Here is the content:

```markdown
src/templates/licenses/ISC_LICENSE.md
<<<<<<< ORIGINAL
=======
ISC License

Copyright (c) [year], [fullname]

Permission to use, copy, modify, and/or distribute this software for any
purpose with or without fee is hereby granted, provided that the above
copyright notice and this permission notice appear in all copies.

THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
>>>>>>> UPDATED
```

Please review this change and let me know if it is correct.